maze2d/requirements.txt
# setuptools==66
absl-py==1.4.0
adal==1.2.7
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
applicationinsights==0.11.10
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
awscli==1.27.62
azure-cli-core==2.11.1
azure-cli-telemetry==1.0.8
azure-common==1.1.28
azure-core==1.26.2
azure-mgmt-authorization==0.61.0
azure-mgmt-compute==13.0.0
azure-mgmt-core==1.2.0
azure-mgmt-network==11.0.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.4.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
boto==2.49.0
boto3==1.26.62
botocore==1.29.62
cachetools==5.3.0
# certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
chex==0.1.5
click==8.0.4
cloudpickle==1.6.0
colorama==0.4.4
comm==0.1.2
contourpy==1.0.7
crcmod==1.7
cryptography==39.0.0
cycler==0.11.0
lockfile==0.12.2
Cython==0.29.33
d4rl @ git+https://github.com/Farama-Foundation/d4rl@f2a05c0d66722499bf8031b094d9af3aea7c372b
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
# -e git+https://github.com/FelipeNuti/diffusion-relative-rewards.git@4f77d31d01b2d331ca31d33c9538fd31dc663e26#egg=diffuser&subdirectory=maze2d
dill==0.3.6
distlib==0.3.6
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.16
# doodad @ git+https://github.com/JannerM/doodad.git@77cdefb9862625bb6f419e66cd502318225dc992
einops==0.6.0
executing==1.2.0
fasteners==0.18
fastjsonschema==2.16.2
ffmpeg==1.4
ffprobe==0.5
filelock==3.9.0
flax==0.4.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.3
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
glfw==2.5.6
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
# gym==0.18.0
h5py==3.8.0
httplib2==0.21.0
humanfriendly==8.2
idna==3.4
imageio==2.9.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
isoduration==20.11.0
jax==0.2.21
jaxlib==0.4.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
knack==0.7.2
labmaze==1.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.4
# mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
msal==1.0.0
msal-extensions==0.1.3
msgpack==1.0.4
msrest==0.7.1
msrestazure==0.6.4
# mujoco==2.3.1.post1
# mujoco-py==2.0.2.13
mypy-extensions==0.4.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
portalocker==1.7.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser==2.21
pyglet==1.5.0
Pygments==2.14.0
PyJWT==1.7.1
PyNaCl==1.5.0
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
# ray==2.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.17.2
scikit-learn==1.2.1
scikit-video==1.1.11
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
scikit-learn==1.2.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tinycss2==1.2.1
toolz==0.12.0
# torch==1.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torch-tb-profiler==0.4.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typed-argument-parser==1.7.2
typing-inspect==0.8.0
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
wandb==0.13.9
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.2.3
widgetsnbextension==4.0.5
zipp==3.12.0
patchelf
wheel==0.38.4
